constraints no longer make wiggles unusable
Wiggle bones seems to specially handle constraints, unfortunately the
math is incorrect and causes bones to twist, offset, and scale inappropriately.
On chains this will immediately cause the simulation to explode.

This is particularly a problem with AutoRigPro rigs, which loves to use
rotation constraints on tails, hair, and ears. (prime wiggle targets!)

With this change, wiggle ignores constraints, which causes odd behaviors
depending on the constraint involved, as constraints have the "final" say
in a pose-- while wiggle operates on an unseen internal pose.

Now instead of exploding when any arbitrary constraint exists on a
bone, it has a much more understandable failure case to a user related
to the involved constraints.

I was unable to test if this broke pinning or not, as pinning as a
feature already doesn't work... So I also took the liberty of removing
the pinning feature.