Optimize Baking by Disabling non-Jiggle objects
At first I tried to create a custom bake funciton, but that broke the physics eval. So I went back to using the NLA method. It seems we can't bypass the NLA bake operation. So let's see if we can make it more efficient.

I added some code to disable objects not related to jiggle, so they don't get evaluated for frame baking.

I just tested, and it is faster, but not significantly. This does mean that its working though. I'll investigate further.