Toggle Button for Debug Bake logging
I added a button on the panel to allow the user to toggle the Baking Log Debugger, since having it on all the time might cause overhead. Now its a simple toggle to turn it on or off, which should help developers and users identify baking bottlenecks in their scene.